requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.0
django-allauth==0.45.0
django-crispy-forms==1.12.0
django-environ==0.4.5
django-summernote==0.8.11.6
idna==3.2
mysqlclient==2.0.3
oauthlib==3.1.1
Pillow==8.2.0
pkg-resources==0.0.0
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.1
urllib3==1.26.6
